Move Models to new project
