+++ Backlogs.Core/Models/SearchResult.cs
